Issue 1 - Moving watch pod to the runtime package.
