+++ apb/provision_or_update.go
@@ -78,7 +78,7 @@ func (e *executor) provisionOrUpdate(method executionMethod, instance *ServiceIn
 
 	if instance.Spec.Runtime >= 2 || !instance.Spec.Bindable {
 		log.Debugf("watching pod for serviceinstance %#v", instance.Spec)
-		err := watchPod(executionContext.PodName, executionContext.Namespace,
+		err := runtime.WatchPod(executionContext.PodName, executionContext.Namespace,
 			k8scli.Client.CoreV1().Pods(executionContext.Namespace), e.updateDescription)
 		if err != nil {
 			log.Errorf("Provision or Update action failed - %v", err)
